Improve the sshnodelay hack after osxfuse's merge
Document the strage-looking symbol interposition that is necessary for the
OS X case.

While doing this, homogenize the code to do the real connect call for both
the Linux and OS X cases into a helper function.  This logic is generic
and should not be replicated into various OS-specific cases.