meson: allow explicitly disabling manpage generation
While manpage generation is optional depending on whether
'rst2man'/'rst2man.py' is found, a broken rst2man will still cause build
breakage.

For a build system like Buildroot, where manpages are anyway not needed,
it is preferred to be able to disable manpage generation altogether, to
avoid any impact of a broken user environment.

Add an option of type 'feature' which can be disabled explicitly as follows:
    -Dbuild-manual=disabled

In that case, even if rst2man would be present, the manpage would still not
be generated.

The 'feature' option type was added in meson 0.47, so bump the minimum meson
version accordingly.

See https://mesonbuild.com/Build-options.html#features .

Signed-off-by: Thomas De Schampheleire <thomas.de_schampheleire@nokia.com>